archive/advanced-live-ai-tool/server/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-dotenv==1.0.1
sqlalchemy==2.0.27
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
redis==5.0.2
pydantic==2.6.1
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
openai==1.12.0
httpx==0.27.0
stripe==8.4.0
websockets==12.0
aiofiles==23.2.1
python-magic==0.4.27
pdfplumber==0.11.0
tiktoken==0.6.0
boto3==1.34.49
celery==5.3.6
flower==2.0.1
sentry-sdk[fastapi]==1.40.4
